+++ models/cutmix.py
@@ -1,0 +1,31 @@
+import torch
+
+
+def rand_bbox(size, lam):
+    W, H = size
+    cut_rat = (1. - lam).sqrt()
+    cut_w = (W * cut_rat).to(torch.long)
+    cut_h = (H * cut_rat).to(torch.long)
+
+    cx = torch.zeros_like(cut_w, dtype=cut_w.dtype).random_(0, W)
+    cy = torch.zeros_like(cut_h, dtype=cut_h.dtype).random_(0, H)
+
+    bbx1 = (cx - cut_w // 2).clamp(0, W)
+    bby1 = (cy - cut_h // 2).clamp(0, H)
+    bbx2 = (cx + cut_w // 2).clamp(0, W)
+    bby2 = (cy + cut_h // 2).clamp(0, H)
+
+    new_lam = 1. - (bbx2 - bbx1).to(lam.dtype) * (bby2 - bby1).to(lam.dtype) / (W * H)
+
+    return (bbx1, bby1, bbx2, bby2), new_lam
+
+
+def cutmix(input, alpha):
+    beta = torch.distributions.beta.Beta(alpha, alpha)
+    randind = torch.randperm(input.shape[0], device=input.device)
+    lam = beta.sample().to(device=input.device)
+    lam = torch.max(lam, 1. - lam)
+    (bbx1, bby1, bbx2, bby2), lam = rand_bbox(input.shape[-2:], lam)
+    output = input.clone()
+    output[..., bbx1:bbx2, bby1:bby2] = output[randind][..., bbx1:bbx2, bby1:bby2]
+    return output, randind, lam
