+++ models/mixup.py
@@ -1,0 +1,11 @@
+import torch
+
+
+def mixup(input, alpha):
+    beta = torch.distributions.beta.Beta(alpha, alpha)
+    randind = torch.randperm(input.shape[0], device=input.device)
+    lam = beta.sample([input.shape[0]]).to(device=input.device)
+    lam = torch.max(lam, 1. - lam)
+    lam_expanded = lam.view([-1] + [1]*(input.dim()-1))
+    output = lam_expanded * input + (1. - lam_expanded) * input[randind]
+    return output, randind, lam
