+++ models/moco_mixup.py
@@ -1,9 +1,10 @@
+from argparse import ArgumentParser
+
 import torch
 from torch.nn import functional as F
 
-import random
-
 from models.moco import MoCo
+from models.mixup import mixup
 from pl_bolts.metrics import precision_at_k
 
 
@@ -11,19 +12,8 @@ class MoCoMixup(MoCo):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
-    def mixup(self, input, alpha):
-        beta = torch.distributions.beta.Beta(alpha, alpha)
-        randind = torch.randperm(input.shape[0], device=input.device)
-        lam = beta.sample([input.shape[0]]).to(device=input.device)
-        lam = torch.max(lam, 1. - lam)
-        lam_expanded = lam.view([-1] + [1]*(input.dim()-1))
-        output = lam_expanded * input + (1. - lam_expanded) * input[randind]
-        return output, randind, lam
-
-    def training_step(self, batch, batch_idx):
-        img1, img2 = self.process_batch(batch)
-
-        img1, target_aux, lam = self.mixup(img1, alpha=1.)
+    def training_step_after_process_batch(self, img1, img2):
+        img1, target_aux, lam = mixup(img1, alpha=self.hparams.alpha)
         target = torch.arange(img1.shape[0], dtype=torch.long).cuda()
 
         self._momentum_update_key_encoder()
@@ -35,11 +25,20 @@ class MoCoMixup(MoCo):
 
         # compute loss
         contrast = torch.cat([k, self.queue.clone().detach()], dim=0)
-        logits = torch.mm(q, contrast.t())
-        loss = lam * F.cross_entropy(logits, target) + (1. - lam) * F.cross_entropy(logits, target_aux)
+        logits = torch.mm(q, contrast.t()) / self.hparams.temperature
+        loss = (lam * F.cross_entropy(logits, target) + (1. - lam) * F.cross_entropy(logits, target_aux)).mean()
         acc1, acc5 = precision_at_k(logits, target, top_k=(1, 5))
 
         self._dequeue_and_enqueue(k)
 
         self.log_dict({'loss': loss, 'acc1': acc1, 'acc5': acc5}, prog_bar=True)
         return loss
+
+    @staticmethod
+    def add_model_specific_args(parent_parser):
+        parent_parser = MoCo.add_model_specific_args(parent_parser)
+        parser = ArgumentParser(parents=[parent_parser], add_help=False)
+
+        # training params
+        parser.add_argument("--alpha", default=1., type=float)
+        return parser
